// Presentation_Angular/ClientApp/src/app/Core/models/registerUser.ts
export interface RegisterUser {
  firstName: string;
  familyName: string;
  username: string;
  email: string;
  password: string;
  confirmPassword: string;
  street: string;
  houseNr: string;
  houseBus: string;
  postalCode: string;
  city: string;
  streetBilling: string;
  houseNrBilling: string;
  houseBusBilling: string;
  postalCodeBilling: string;
  cityBilling: string;
}